+++ LearnRLSK/__main__.py
@@ -1,272 +1,5 @@
-from LearnRLSK.RLframe import *
-from LearnRLSK.utilities import *
-import numpy as np
-import warnings
-import sys
+from LearnRLSK import Simulation
 import argparse
-
-
-class Simulation:
-    """class to create simulation and run simulation."""
-
-    def __init__(self,
-                 dim_state=5,
-                 dim_input=2,
-                 dim_output=5,
-                 dimDisturb=2,
-                 m=10,
-                 I=1,
-                 t0=0,
-                 t1=100,
-                 n_runs=1,
-                 a_tol=1e-5,
-                 r_tol=1e-3,
-                 x_min=-10,
-                 x_max=10,
-                 y_min=-10,
-                 y_max=10,
-                 dt=0.05,
-                 mod_est_phase=2,
-                 model_order=5,
-                 prob_noise_pow=8,
-                 mod_est_checks=0,
-                 f_man=-3,
-                 n_man=-1,
-                 f_min=-5,
-                 f_max=5,
-                 m_min=-1,
-                 m_max=1,
-                 nactor=6,
-                 buffer_size=200,
-                 r_cost_struct=1,
-                 n_critic=50,
-                 gamma=1,
-                 critic_struct=3,
-                 is_log_data=0,
-                 is_visualization=1,
-                 is_print_sim_step=1,
-                 is_disturb=0,
-                 is_dyn_ctrl=0,
-                 ctrl_mode=5):
-        """init."""
-        self.dim_state = dim_state
-        self.dim_input = dim_input
-        self.dim_output = dim_output
-        self.dimDisturb = dimDisturb
-        self.m = m
-        self.I = I
-        self.sigma_q = 1e-3 * np.ones(dimDisturb)
-        self.mu_q = np.zeros(dimDisturb)
-        self.tau_q = np.ones(dimDisturb)
-        self.t0 = t0
-        self.t1 = t1
-        self.n_runs = n_runs
-        self.x0 = np.zeros(dim_state)
-        self.x0[0] = 5
-        self.x0[1] = 5
-        self.x0[2] = np.pi / 2
-        self.u0 = 0 * np.ones(dim_input)
-        self.q0 = 0 * np.ones(dimDisturb)
-        self.a_tol = a_tol
-        self.r_tol = r_tol
-        self.x_min = x_min
-        self.x_max = x_max
-        self.y_min = y_min
-        self.y_max = y_max
-        self.dt = dt
-        self.mod_est_phase = mod_est_phase
-        self.modEstPeriod = 1 * dt
-        self.model_order = model_order
-        self.prob_noise_pow = prob_noise_pow
-        self.mod_est_checks = mod_est_checks
-        self.f_man = f_man
-        self.n_man = n_man
-        self.uMan = np.array([f_man, n_man])
-        self.f_min = f_min
-        self.f_max = f_max
-        self.m_min = m_min
-        self.m_max = m_max
-        self.nactor = nactor
-        self.predStepSize = 5 * dt
-        self.buffer_size = buffer_size
-        self.r_cost_struct = r_cost_struct
-        self.R1 = np.diag([10, 10, 1, 0, 0, 0, 0])
-        self.R2 = np.array([[10, 2, 1, 0, 0],
-                            [0, 10, 2, 0, 0],
-                            [0, 0, 10, 0, 0],
-                            [0, 0, 0, 0, 0],
-                            [0, 0, 0, 0, 0]])
-        self.n_critic = n_critic
-        self.gamma = gamma
-        self.critic_period = 5 * dt
-        self.critic_struct = critic_struct
-        self.is_log_data = is_log_data
-        self.is_visualization = is_visualization
-        self.is_print_sim_step = is_print_sim_step
-        self.is_disturb = is_disturb
-        self.is_dyn_ctrl = is_dyn_ctrl
-        self.ctrl_mode = ctrl_mode
-
-    def run_sim(self):
-        """run sim."""
-        ctrlBnds = np.array([[self.f_min, self.f_max], [self.m_min, self.m_max]])
-
-        # environment
-        sys = system(self.dim_state,
-                     self.dim_input,
-                     self.dim_output,
-                     self.dimDisturb,
-                     pars=[self.m, self.I],
-                     ctrlBnds=ctrlBnds)
-
-        alpha0 = self.x0[2]
-
-        # agent
-        myNominalCtrl = nominalController(self.m,
-                                          self.I,
-                                          ctrlGain=0.5,
-                                          ctrlBnds=ctrlBnds,
-                                          t0=self.t0,
-                                          samplTime=self.dt)
-
-        agent = controller(self.dim_input,
-                           self.dim_output,
-                           self.ctrl_mode,
-                           ctrlBnds=ctrlBnds,
-                           t0=self.t0,
-                           samplTime=self.dt,
-                           Nactor=self.nactor,
-                           predStepSize=self.predStepSize,
-                           sysRHS=sys._stateDyn,
-                           sysOut=sys.out,
-                           xSys=self.x0,
-                           probNoisePow=self.prob_noise_pow,
-                           modEstPhase=self.mod_est_phase,
-                           modEstPeriod=self.modEstPeriod,
-                           bufferSize=self.buffer_size,
-                           modelOrder=self.model_order,
-                           modEstChecks=self.mod_est_checks,
-                           gamma=self.gamma,
-                           Ncritic=self.n_critic,
-                           criticPeriod=self.critic_period,
-                           criticStruct=self.critic_struct,
-                           rcostStruct=self.r_cost_struct,
-                           rcostPars=[self.R1, self.R2])
-
-        # simulator
-        if self.is_dyn_ctrl:
-            ksi0 = np.concatenate([self.x0, self.q0, self.u0])
-        else:
-            ksi0 = np.concatenate([self.x0, self.q0])
-
-        simulator = sp.integrate.RK45(sys.closedLoop,
-                                      self.t0, ksi0, self.t1,
-                                      max_step=self.dt / 2,
-                                      first_step=1e-6,
-                                      atol=self.a_tol,
-                                      rtol=self.r_tol)
-
-        # extras
-        dataFiles = logdata(self.n_runs, save=self.is_log_data)
-
-        if self.is_print_sim_step:
-            warnings.filterwarnings('ignore')
-
-        # main loop
-
-        if self.is_visualization:
-            myAnimator = animator(objects=(simulator,
-                                           sys,
-                                           myNominalCtrl,
-                                           agent,
-                                           dataFiles,
-                                           ctrlSelector,
-                                           printSimStep,
-                                           logDataRow),
-                                  pars=(self.dim_state,
-                                        self.x0,
-                                        self.u0,
-                                        self.t0,
-                                        self.t1,
-                                        ksi0,
-                                        self.x_min,
-                                        self.x_max,
-                                        self.y_min,
-                                        self.y_max,
-                                        self.f_min,
-                                        self.f_max,
-                                        self.m_min,
-                                        self.m_max,
-                                        self.ctrl_mode,
-                                        self.uMan,
-                                        self.n_runs,
-                                        self.is_print_sim_step,
-                                        self.is_log_data))
-
-            anm = animation.FuncAnimation(myAnimator.simFig,
-                                          myAnimator.animate,
-                                          init_func=myAnimator.initAnim,
-                                          blit=False,
-                                          interval=self.dt / 1e6,
-                                          repeat=True)
-
-            anm.running = True
-            myAnimator.simFig.canvas.mpl_connect('key_press_event',
-                                                 lambda event: onKeyPress(event, anm))
-            myAnimator.simFig.tight_layout()
-            plt.show()
-
-        else:
-            currRun = 1
-            dataFile = dataFiles[0]
-
-            while True:
-                simulator.step()
-                t = simulator.t
-                ksi = simulator.y
-                x = ksi[0:dim_state]
-                y = sys.out(x)
-                u = ctrlSelector(
-                    t, y, self.uMan, myNominalCtrl, agent, ctrl_mode)
-                sys.receiveAction(u)
-                agent.receiveSysState(sys._x)
-                agent.update_icost(y, u)
-                xCoord = ksi[0]
-                yCoord = ksi[1]
-                alpha = ksi[2]
-                v = ksi[3]
-                omega = ksi[4]
-                icost = agent.icostVal
-
-                if self.is_print_sim_step:
-                    printSimStep(t, xCoord, yCoord, alpha, v, omega, icost, u)
-
-                if self.is_log_data:
-                    logDataRow(dataFile, t, xCoord, yCoord,
-                               alpha, v, omega, icost, u)
-
-                if t >= self.t1:
-                    if is_print_sim_step:
-                        print('.....................................Run {run:2d} done.....................................'.format(
-                            run=currRun))
-                    currRun += 1
-                    if currRun > n_runs:
-                        break
-
-                    if self.is_log_data:
-                        dataFile = dataFiles[currRun - 1]
-
-                    # Reset simulator
-                    simulator.status = 'running'
-                    simulator.t = t0
-                    simulator.y = ksi0
-
-                    if ctrl_mode > 0:
-                        agent.reset(self.t0)
-                    else:
-                        myNominalCtrl.reset(self.t0)
-                    icost = 0
-
 
 def main(args=None):
     """main."""
